Correct geolocation data types, small improvements

+++ plugins/weather/src/main/java/de/unistuttgart/iaas/amyassist/amy/plugin/weather/WeatherLogic.java
@@ -81,7 +81,7 @@ public class WeatherLogic {
 		 * @param pLatitude
 		 *            The latitude to use
 		 */
-		public GeoCoordinatePair(long pLongitude, long pLatitude) {
+		public GeoCoordinatePair(double pLongitude, double pLatitude) {
 			this.longitude = pLongitude;
 			this.latitude = pLatitude;
 		}
